GEOMESA-2937 Fix schema management index flag checks (#2640)
* GEOMESA-2937 Fix schema management index flag checks

* Split up GeoMesaDataStore test classes
* Remove accidental check-in of application conf file
