--- geomesa-tools/conf/application2.conf
@@ -1,35 +1,0 @@
-geomesa {
-  sfts {
-    "example-csv" = {
-      attributes = [
-        { name = "fid",      type = "Integer",         index = false }
-        { name = "name",     type = "String",          index = true  }
-        { name = "age",      type = "Integer",         index = false }
-        { name = "lastseen", type = "Date",            index = false, default = true }
-        { name = "geom",     type = "Point",           index = true, srid = 4326, default = true }
-      ]
-    }
-  }
-}
-
-geomesa {
-  converters {
-    "example-csv" = {
-      type   = "delimited-text"
-      format = "CSV"
-      options {
-        skip-lines = 1
-      }
-      id-field = "uuid()",
-      fields = [
-        { name = "fid",      transform = "$1::int"                     }
-        { name = "name",     transform = "$2::string"                  }
-        { name = "age",      transform = "$3::int"                     }
-        { name = "lastseen", transform = "date('yyyy-MM-dd', $4)"      }
-        { name = "lon",      transform = "$7::double"                  }
-        { name = "lat",      transform = "$8::double"                  }
-        { name = "geom",     transform = "point($lon, $lat)"           }
-      ]
-    }
-  }
-}
